at32f4: rcc: split out less common code
