+++ lib/at32/f40x/Makefile
@@ -43,7 +43,7 @@ OBJS		+= gpio.o gpio_common_all.o
 OBJS		+= i2c_common_v1.o
 OBJS		+= iwdg_common_all.o
 OBJS		+= pwr_common_v1.o
-OBJS		+= rcc.o rcc_common_at32.o rcc_common_all.o
+OBJS		+= rcc.o rcc_common_f02.o rcc_common_at32.o rcc_common_all.o
 OBJS		+= rtc.o
 OBJS		+= spi_common_all.o spi_common_v1.o
 OBJS		+= timer_common_at32.o timer_common_all.o
