+++ lib/at32/f42x/Makefile
@@ -41,7 +41,7 @@ OBJS		+= gpio_common_all.o gpio_common_f0234.o gpio_common_at32.o
 OBJS		+= i2c_common_v2.o
 OBJS		+= iwdg_common_all.o
 OBJS		+= pwr_common_v1.o
-OBJS		+= rcc.o rcc_common_at32.o rcc_common_all.o
+OBJS		+= rcc.o rcc_common_f02.o rcc_common_at32.o rcc_common_all.o
 OBJS		+= rtc_common_l1f024.o
 OBJS		+= spi_common_all.o spi_common_v1.o spi_common_v1_frf.o
 OBJS		+= timer_common_at32.o timer_common_all.o
